colorization_project/requirements.txt
# 项目依赖

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 图像处理
scikit-image>=0.20.0
opencv-python>=4.7.0
Pillow>=9.5.0

# 数据处理
numpy>=1.24.0
scipy>=1.10.0

# 可视化和日志
matplotlib>=3.7.0
tensorboard>=2.12.0
tqdm>=4.65.0

# 评估指标
scikit-learn>=1.2.0

# 工具
pyyaml>=6.0
